refactor: Modularize socket message event handlers
- remove `Chat` and `Classroom` model integration and event handling logic in `messageController.js`.
- integrate `Chat` and `Classroom`  models in `socket.messages.js`.
- implement ,`sendMessage` `handleChatOpen`, and `handleTyping`  functions with socket events `message room`, `user typing` and `open conversation` handling logic in `socket.messages.js`.
- integrate `sendMessage`, `handleChatOpen`, and `handlTyping` in matching event handlers, and remove in `messageController.js`.
